add rule selection
users may pass a function with accepts an iterator over reductions, which contain a new tree and a private field containing the associated side-effects of the reduction. Through this the user can provide arbitrary rule selection logic, like always choosing the first applicable rule, choosing a random one, etc.

Since a new Commands object must be created for each reduction, it is no longer enough for one mutable Commands object to be created at the top of the tree and passed down. Instead, the commands object to apply is passed back up the tree alongside the rewriting process. This means that internal engine functions now use Reductions under the hood as well.